refactor: TestUtils에서 MvcResult 응답 객체 받는다
